// src/common/enums/authentication.ts
export enum UserGroup {
	ADMINISTRATOR = "EVOS-ADMINISTRATOR",
	VOICE_ENGINEER = "EVOS-VOICE_ENGINEER",
	VENDOR = "EVOS-VENDOR",
	READ_ONLY = "EVOS-READ_ONLY",
}

export enum AppRole {
	SERVICE_PROVIDER_ATT_READ = "ServiceProvider.AT&T.Read",
	SERVICE_PROVIDER_ATT_WRITE = "ServiceProvider.AT&T.Write",
	SERVICE_PROVIDER_VERIZON_READ = "ServiceProvider.Verizon.Read",
	SERVICE_PROVIDER_VERIZON_WRITE = "ServiceProvider.Verizon.Write",
}